feat(2.12): support 2.12 compiler
Fixing broken links in CONTRIBUTING.md while I'm in here.

Using a matrix style build to ensure we publish w/ the lowest JDK for each
Scala version. Idea is that if we use the current scala version when +publish, then
we should pickup whatever travis is using from our matrix and get e.g.,
_2.11 w/ JDK7 and _2.12 w/ JDK8. Should help w/ binary compat and
simplify the build, but leaving as snapshot to test it out.